crawler-service/requirements.txt
scrapy
scrapy-playwright
redis
beautifulsoup4
rake-nltk
spacy
pydantic
python-dotenv
